requirements.txt
aenum==3.1.11
affine==2.3.1
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudinary==1.29.0
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
dataclasses-json==0.5.7
dj-database-url==0.5.0
Django==4.0.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
fonttools==4.34.4
GDAL @ file:///E:/GDAL-3.4.3-cp310-cp310-win_amd64.whl
graphviz==0.20
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imageio==2.20.0
kiwisolver==1.4.4
marshmallow==3.17.0
marshmallow-enum==1.5.1
matplotlib==3.5.2
mypy-extensions==0.4.3
numpy==1.23.1
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproj==3.3.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
rasterio==1.3.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.3.0
sentinelhub==3.6.3
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
sqlparse==0.4.2
tifffile==2022.5.4
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing-inspect==0.7.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
utm==0.7.0
whitenoise==6.2.0
wsproto==1.1.0
